requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
docutils==0.21.2
idna==3.7
Kivy==2.3.0
Kivy-Garden==0.1.5
kivymd==1.2.0
numpy==1.26.4
opencv-python==4.10.0.82
pillow==10.3.0
Pygments==2.18.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
requests==2.32.3
urllib3==2.2.1

